# src/Recruit.Api/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Recruit.Api",
    "version": "v1"
  },
  "paths": {
...
    "/api/user/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Version",
            "in": "header",
            "required": true,
            "example": "1.0"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RecruitUser"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecruitUser"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecruitUser"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Version",
            "in": "header",
            "required": true,
            "example": "1.0"
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/PutUserRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PutUserRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PutUserRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PutUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PutUserResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PutUserResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PutUserResponse"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PutUserResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PutUserResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PutUserResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Version",
            "in": "header",
            "required": true,
            "example": "1.0"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Operation"
                }
              }
            },
            "application/json-patch+json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Operation"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Operation"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Operation"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RecruitUser"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecruitUser"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecruitUser"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/by/email": {
      "post": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "X-Version",
            "in": "header",
            "required": true,
            "example": "1.0"
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/GetUserRequest"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetUserRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/GetUserRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/GetUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RecruitUser"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecruitUser"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecruitUser"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/by/employeraccountid/{employerAccountId}": {
      "get": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "employerAccountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "notificationType",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/NotificationTypes"
            }
          },
          {
            "name": "X-Version",
            "in": "header",
            "required": true,
            "example": "1.0"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecruitUser"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecruitUser"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecruitUser"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/user/by/ukprn/{ukprn}": {
      "get": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "ukprn",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "notificationType",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/NotificationTypes"
            }
          },
          {
            "name": "X-Version",
            "in": "header",
            "required": true,
            "example": "1.0"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecruitUser"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecruitUser"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecruitUser"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/user/by/dfeuserid/{dfeUserId}": {
      "get": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "dfeUserId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Version",
            "in": "header",
            "required": true,
            "example": "1.0"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RecruitUser"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecruitUser"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecruitUser"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/by/idams/{idams}": {
      "get": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "idams",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "X-Version",
            "in": "header",
            "required": true,
            "example": "1.0"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RecruitUser"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecruitUser"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecruitUser"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/vacancies/count/user/{userId}": {
      "get": {
        "tags": [
          "Vacancy"
        ],
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Version",
            "in": "header",
            "required": true,
            "example": "1.0"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Int32DataResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Int32DataResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Int32DataResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "GetUserRequest": {
        "required": [
          "email",
          "userType"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "nullable": true
          },
          "userType": {
            "$ref": "#/components/schemas/UserType"
          }
        },
        "additionalProperties": false
      },
      "Int32DataResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "NotificationFrequency": {
        "enum": [
          "NotSet",
          "Never",
          "Immediately",
          "Daily",
          "Weekly"
        ],
        "type": "string"
      },
      "NotificationPreference": {
        "type": "object",
        "properties": {
          "event": {
            "$ref": "#/components/schemas/NotificationTypes"
          },
          "method": {
            "type": "string",
            "nullable": true
          },
          "scope": {
            "$ref": "#/components/schemas/NotificationScope"
          },
          "frequency": {
            "$ref": "#/components/schemas/NotificationFrequency"
          }
        },
        "additionalProperties": false
      },
      "NotificationPreferences": {
        "type": "object",
        "properties": {
          "eventPreferences": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NotificationPreference"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "NotificationScope": {
        "enum": [
          "NotSet",
          "UserSubmittedVacancies",
          "OrganisationVacancies"
        ],
        "type": "string"
      },
      "NotificationTypes": {
        "enum": [
          "VacancyApprovedOrRejected",
          "VacancyClosingSoon",
          "ApplicationSubmitted",
          "VacancySentForReview",
          "SharedApplicationReviewedByEmployer",
          "ProviderAttachedToVacancy",
          "ApplicationSharedWithEmployer"
        ],
        "type": "string"
      },
      "Operation": {
        "type": "object",
        "properties": {
          "path": {
            "type": "string",
            "nullable": true
          },
          "op": {
            "type": "string",
            "nullable": true
          },
          "from": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "PutUserRequest": {
        "required": [
          "email",
          "name",
          "userType"
        ],
        "type": "object",
        "properties": {
          "idamsUserId": {
            "type": "string",
            "nullable": true
          },
          "userType": {
            "$ref": "#/components/schemas/UserType"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          },
          "email": {
            "minLength": 1,
            "type": "string"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "lastSignedInDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "employerAccountIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            },
            "nullable": true
          },
          "ukprn": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "transferredVacanciesEmployerRevokedPermissionAlertDismissedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "closedVacanciesBlockedProviderAlertDismissedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "transferredVacanciesBlockedProviderAlertDismissedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "closedVacanciesWithdrawnByQaAlertDismissedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "notificationPreferences": {
            "$ref": "#/components/schemas/NotificationPreferences"
          },
          "dfEUserId": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PutUserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "additionalProperties": false
      },
...
      "RecruitUser": {
        "required": [
          "email",
          "name",
          "userType"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "idamsUserId": {
            "maxLength": 200,
            "type": "string",
            "nullable": true
          },
          "dfEUserId": {
            "maxLength": 200,
            "type": "string",
            "nullable": true
          },
          "userType": {
            "$ref": "#/components/schemas/UserType"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "updatedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastSignedInDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "name": {
            "maxLength": 500,
            "type": "string",
            "nullable": true
          },
          "email": {
            "maxLength": 250,
            "type": "string",
            "nullable": true
          },
          "employerAccountIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int64"
            },
            "nullable": true
          },
          "ukprn": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "transferredVacanciesEmployerRevokedPermissionAlertDismissedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "closedVacanciesBlockedProviderAlertDismissedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "transferredVacanciesBlockedProviderAlertDismissedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "closedVacanciesWithdrawnByQaAlertDismissedOn": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "notificationPreferences": {
            "$ref": "#/components/schemas/NotificationPreferences"
          }
        },
        "additionalProperties": false
      },
...
      "UserType": {
        "enum": [
          "Employer",
          "Provider"
        ],
        "type": "string"
      },
...
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
    }
  }
}